// apps/frontend/core-dapp/src/types/market.types.ts
export type Range = [number, number]

export type SupportedMarketID = 'arbitrum' | 'zksync' | 'starknet'

export type MarketType = 'preICO' | 'preIPO'

export type MarketVersion = 'v1.0' | 'v1.1'

export type Market = {
  iconName: IconName
  type: MarketType
  name: string
  urlId: SupportedMarketID
  companyName: string
  totalSupply?: number
  address: SupportedMarkets
  valuationMultiplier?: number
  hideExpiry?: boolean
  long: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  short: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  static: {
    valuationRange: Range
  }
  version: MarketVersion
  settlementDocsLink?: string
}

export type MarketHistoryData = {
  timestamp: number
  volume: number
  valuation: number
  liquidity?: number
}

export type SeparatedMarketHistoryData = {
  long: MarketHistoryData[]
  short: MarketHistoryData[]
}

/**
 * @deprecated just use `number` directly instead
 */
export type NumberData = {
  value: number
}
// ...
export type NormalizedTokenPairWithoutPrice = {
  token0: Omit<NormalizedToken, 'price'>
  token1: Omit<NormalizedToken, 'price'>
}

export type NormalizedTokenPairWithTvlOnly = {
  token0: Pick<NormalizedToken, 'id' | 'volume'>
  token1: Pick<NormalizedToken, 'id' | 'volume'>
}

export type NormalizedTokenPairWithVolumeOnly = {
  token0: Pick<NormalizedToken, 'id' | 'volume'>
  token1: Pick<NormalizedToken, 'id' | 'volume'>
}

export type NormalizedPoolData = { liquidity: number } & NormalizedTokenPairWithoutPrice

export type NormalizedPoolsData = {
  longTokenPool: NormalizedPoolData
  shortTokenPool: NormalizedPoolData
}